Implemented Group DM Creation and Modification
I added a method for creating new group DMs to JDAClient and created a
GroupManager for renaming group DMs and setting their icons because I
needed these features for a project. Small changes were made to
Requester.java and Route.java thanks to a wierd Discord API glitch where
I needed to include 'v6' in the route for Group DM creation to succeed.
The alternate URL is only used in a special route I added specifically
for the task of creating group DMs so my changes aren't going to effect
existing code.