Changed the way the body is parsed even further
This also allows for other types than json object/array and string to be parsed out of the body.
Only difference in behavior is that this approach will throw an exception instead of returning null when the parsing fails. This shouldn't make much of a difference though as the returned object has never been checked for null anyway.