Improve gradle builds and debugging (#674)
We now use the java-library plugin in favor of a better
application binary interface. This will allow use to specify which
dependencies we want to keep in our frontend and which are just for
internal use in JDA.

As such the dependencies added with "implementation" will only be
for runtime use and are not usable for library extensions directly.
This will require libraries using JDA to specify our internal
dependencies explicitly rather than relying on us to keep them, allowing
us to remove dependencies without breaking compilation of other tools.