Remove pom type from pom dependency because gradle
Apparently gradle thinks it makes sense to put pom dependencies
on the classpath when it can't find a jar for it. Beats me.
I am testing this to see if anyone will run into more severe
issues or whether this resolves it completely.
The dependency is using pom packaging so I am hoping
that maven can understand it without the type.